remove or silence warnings about narrowing conversions, unused variables, loss of const-qualifications, and removed unreferenced non-public local functions (msvc)
(cherry picked from commit ef9b58b610a5aacd47d82cd8a90493401b4aeba3)

+++ test/cmdline_test.cpp
@@ -48,7 +48,7 @@ int translate_syntax_error_kind(invalid_command_line_syntax::kind_t k)
     e = table + sizeof(table)/sizeof(table[0]);
     i = std::find(b, e, k);
     assert(i != e);
-    return std::distance(b, i) + 3;
+    return static_cast<int>(std::distance(b, i)) + 3;
 }
 
 struct test_case {
